Switching to PEDANTIC compilation and fixing all warnings
git-svn-id: svn+ssh://svn.cern.ch/reps/IfaeAnaRepo/IFAEanalysis/RunIITopTools/IFAETopFramework/trunk@10840 4525493e-7705-40b1-a816-d608a930855b

+++ Root/OutputManager.cxx
@@ -224,7 +224,7 @@ bool OutputManager::FillTH1FromVector( void* t, const VariableDef::VariableType 
     
     if(type == VariableDef::VECDOUBLE){
         std::vector < double >* vec = (std::vector<double>*)t;
-        if(index!=-1 && vec->size()>index){
+        if(index!=-1 && (int)vec->size()>index){
             m_histMngr -> FillTH1D((std::string)histName, vec->at(index), weight);
         } else if (index==-1){
             for ( double value : *vec ){
@@ -234,7 +234,7 @@ bool OutputManager::FillTH1FromVector( void* t, const VariableDef::VariableType 
     }
     else if(type == VariableDef::VECFLOAT){
         std::vector < float >* vec = (std::vector< float >*)t;
-        if(index!=-1 && vec->size()>index){
+        if(index!=-1 && (int)vec->size()>index){
             m_histMngr -> FillTH1D((std::string)histName, (double)vec->at(index), weight);
         } else if (index==-1){
             for ( double value : *vec ){
@@ -244,7 +244,7 @@ bool OutputManager::FillTH1FromVector( void* t, const VariableDef::VariableType 
     }
     else if(type == VariableDef::VECINT){
         std::vector < int >* vec = (std::vector< int >*)t;
-        if(index!=-1 && vec->size()>index){
+        if(index!=-1 && (int)vec->size()>index){
             m_histMngr -> FillTH1D((std::string)histName, (double)vec->at(index), weight);
         } else if (index==-1){
             for ( double value : *vec ){
